# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ingestor System API
  description: |-
    RESTful API for the Ingestor System that provides access to content processing,
    entity extraction, and database capabilities using Claude AI.
  version: 1.0.0
  contact:
    email: support@example.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/api/v1'
    description: Local development server
...
paths:
...
  /entities:
    get:
      tags:
        - entities
      summary: List all entities
      description: Returns a paginated list of entities with optional filtering
      operationId: listEntities
      parameters:
        - name: type
          in: query
          description: Filter by entity type
          schema:
            type: string
            example: person
        - name: q
          in: query
          description: Search query for entity name or properties
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: sort
          in: query
          description: Field to sort by
          schema:
            type: string
            default: createdAt
            enum: [createdAt, name, type, confidence]
        - name: order
          in: query
          description: Sort order
          schema:
            type: string
            default: desc
            enum: [asc, desc]
        - name: minConfidence
          in: query
          description: Minimum confidence score for entities
          schema:
            type: number
            format: float
            minimum: 0
            maximum: 1
            default: 0.5
      responses:
        '200':
          description: List of entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Entity'
                  meta:
                    type: object
                    properties:
                      pagination:
                        type: object
                        properties:
                          limit:
                            type: integer
                            example: 50
                          offset:
                            type: integer
                            example: 0
                          total:
                            type: integer
                            example: 342
                          next:
                            type: string
                            example: /api/v1/entities?limit=50&offset=50
                            nullable: true
                          previous:
                            type: string
                            example: null
                            nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    post:
      tags:
        - entities
      summary: Create entity
      description: Creates a new entity manually
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityCreate'
      responses:
        '201':
          description: Entity created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Entity'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/types:
    get:
      tags:
        - entities
      summary: Get entity types
      description: Returns all available entity types with their counts
      operationId: getEntityTypes
      responses:
        '200':
          description: List of entity types
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          example: person
                        count:
                          type: integer
                          example: 125
                        description:
                          type: string
                          example: Individual person or character
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/{id}:
    get:
      tags:
        - entities
      summary: Get entity
      description: Returns a specific entity by ID
      operationId: getEntity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entity details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Entity'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    put:
      tags:
        - entities
      summary: Update entity
      description: Updates an existing entity
      operationId: updateEntity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntityUpdate'
      responses:
        '200':
          description: Entity updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Entity'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
    
    delete:
      tags:
        - entities
      summary: Delete entity
      description: Deletes an entity
      operationId: deleteEntity
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entity deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Entity with ID 123e4567-e89b-12d3-a456-426614174000 deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/{id}/related:
    get:
      tags:
        - entities
      summary: Get related entities
      description: Returns entities related to the specified entity
      operationId: getRelatedEntities
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          description: Filter by relationship type
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of related entities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        entity:
                          $ref: '#/components/schemas/Entity'
                        relationship:
                          type: string
                          example: worksFor
                        direction:
                          type: string
                          enum: [outgoing, incoming]
                          example: outgoing
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/{id}/content:
    get:
      tags:
        - entities
      summary: Get entity content
      description: Returns content items associated with the specified entity
      operationId: getEntityContent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Pagination offset
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of content items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/relationships:
    post:
      tags:
        - entities
      summary: Create relationship
      description: Creates a relationship between two entities
      operationId: createRelationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sourceId
                - targetId
                - type
              properties:
                sourceId:
                  type: string
                  format: uuid
                  example: 123e4567-e89b-12d3-a456-426614174000
                targetId:
                  type: string
                  format: uuid
                  example: 987e6543-e89b-12d3-a456-426614174000
                type:
                  type: string
                  example: worksFor
                properties:
                  type: object
                  description: Optional additional properties for the relationship
                  example:
                    startDate: "2022-05-10"
                    role: "Engineer"
      responses:
        '201':
          description: Relationship created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: 456e7890-f12a-34b5-c678-789012345678
                      sourceId:
                        type: string
                        format: uuid
                        example: 123e4567-e89b-12d3-a456-426614174000
                      targetId:
                        type: string
                        format: uuid
                        example: 987e6543-e89b-12d3-a456-426614174000
                      type:
                        type: string
                        example: worksFor
                      properties:
                        type: object
                        example:
                          startDate: "2022-05-10"
                          role: "Engineer"
                      createdAt:
                        type: string
                        format: date-time
                        example: '2025-05-14T04:21:56Z'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Relationship already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  /entities/relationships/{id}:
    delete:
      tags:
        - entities
      summary: Delete relationship
      description: Deletes a relationship between entities
      operationId: deleteRelationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Relationship deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: Relationship with ID 456e7890-f12a-34b5-c678-789012345678 deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Relationship not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
  
  #
  # Batch Processing Endpoints
  #
...
components:
  schemas:
...
    ContentListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/ContentResponse/properties/data'
        meta:
          type: object
          properties:
            pagination:
              type: object
              properties:
                limit:
                  type: integer
                  example: 20
                offset:
                  type: integer
                  example: 0
                total:
                  type: integer
                  example: 156
                next:
                  type: string
                  example: /api/v1/content?limit=20&offset=20
                  nullable: true
                previous:
                  type: string
                  example: null
                  nullable: true
    Entity:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 123e4567-e89b-12d3-a456-426614174000
        type:
          type: string
          example: person
        name:
          type: string
          example: Claude
        properties:
          type: object
          properties:
            description:
              type: string
              example: AI assistant
            organization:
              type: string
              example: Anthropic
        confidence:
          type: number
          format: float
          minimum: 0
          maximum: 1
          example: 0.95
        source:
          type: string
          enum: [extraction, manual, import]
          example: extraction
        createdAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:10Z'
        updatedAt:
          type: string
          format: date-time
          example: '2025-05-14T04:22:10Z'
        contentCount:
          type: integer
          example: 5
          description: Number of content items where this entity appears
        links:
          type: object
          properties:
            self:
              type: string
              example: /api/v1/entities/123e4567-e89b-12d3-a456-426614174000
            related:
              type: string
              example: /api/v1/entities/123e4567-e89b-12d3-a456-426614174000/related
            content:
              type: string
              example: /api/v1/entities/123e4567-e89b-12d3-a456-426614174000/content
    EntityCreate:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          example: Claude
        type:
          type: string
          example: person
        properties:
          type: object
          properties:
            description:
              type: string
              example: AI assistant
            organization:
              type: string
              example: Anthropic
            customProperties:
              type: object
              additionalProperties: true
              example:
                role: Assistant
                language: English
        contentIds:
          type: array
          description: IDs of content items to associate with this entity
          items:
            type: string
            format: uuid
          example: ["123e4567-e89b-12d3-a456-426614174001", "123e4567-e89b-12d3-a456-426614174002"]
    EntityUpdate:
      type: object
      properties:
        name:
          type: string
          example: Claude AI
        type:
          type: string
          example: person
        properties:
          type: object
          properties:
            description:
              type: string
              example: Advanced AI assistant
            organization:
              type: string
              example: Anthropic
            customProperties:
              type: object
              additionalProperties: true
              example:
                role: Advanced Assistant
                language: English
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation error
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                    example: content
                  message:
                    type: string
                    example: Content is required
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key